refactor(e2e): extract the shared cluster-free harness server setup
playwright.legacy-local.config.ts inlined the backend webServer entry, the
app-config stack and the repo-root .bin PATH shim. A second cluster-free config
needs the same backend boot with a different overlay, so move the shared pieces
to playwright/support/local-harness-servers.ts and have the legacy config
consume them. No behaviour change: the extracted backendWebServer() is
equivalent to the inlined entry it replaces.
