+++ e2e-tests/playwright.legacy-local.config.ts
@@ -1,6 +1,11 @@
-import { resolve } from "path";
-
 import { defineConfig, devices } from "@playwright/test";
+
+import {
+  backendWebServer,
+  harnessConfigArgs,
+  isCI,
+  pathWithRepoBin,
+} from "./playwright/support/local-harness-servers";
 
 // Packaged-app sidebar markup differs from the RHDH container global-header sidebar
 // even though /healthcheck returns JSON via the dev-server proxy.
@@ -22,27 +27,13 @@ process.env.E2E_FORCE_LEGACY_SIDEBAR = "true";
  *   yarn --cwd e2e-tests e2e:legacy-local
  *
  * Both servers are started via `webServer` with the guest-auth overlay
- * `app-config.local-e2e.yaml` plus the dynamic-plugins UI config. An already-running
- * pair of servers is reused locally; in CI they are started fresh.
+ * `app-config.local-e2e.yaml` plus the dynamic-plugins UI config (shared setup:
+ * playwright/support/local-harness-servers.ts). An already-running pair of
+ * servers is reused locally; in CI they are started fresh. Re-run populate.sh
+ * after changing the harness plugin set.
  */
 
 const frontendUrl = "http://localhost:3000";
-const backendReadiness = "http://localhost:7007/.backstage/health/v1/readiness";
-const repoRootBin = resolve(process.cwd(), "..", "node_modules", ".bin");
-const pathWithRepoBin = `${repoRootBin}:${process.env.PATH ?? ""}`;
-const isCI = process.env.CI !== undefined && process.env.CI !== "";
-
-// The last entry is generated by e2e-tests/local-harness/populate.sh: it carries the
-// pluginConfig blocks from the harness dynamic-plugins.yaml (e.g. the full global-header
-// mount points), merged by install-dynamic-plugins exactly like the production container
-// does. It must come after the static app-config.dynamic-plugins.yaml so its plugin
-// config wins. Re-run populate.sh after changing the harness plugin set.
-const sharedConfigArgs = [
-  "--config ../../app-config.yaml",
-  "--config ../../app-config.dynamic-plugins.yaml",
-  "--config ../../app-config.local-e2e.yaml",
-  "--config ../../dynamic-plugins-root/app-config.dynamic-plugins.yaml",
-].join(" ");
 
 export default defineConfig({
   testDir: "./playwright",
@@ -103,27 +94,10 @@ export default defineConfig({
   expect: {
     timeout: 15 * 1000,
   },
-  // backstage-cli / janus-cli live in the repo-root node_modules/.bin, which yarn does
-  // not surface for these workspaces, so both CLIs are invoked directly with the root
-  // .bin prepended to PATH and run from their package directory. The backend command
-  // mirrors packages/backend's `start` script (--require instrumentation) — keep in sync.
   webServer: [
+    backendWebServer(),
     {
-      command: `backstage-cli package start --require ./src/instrumentation.js ${sharedConfigArgs}`,
-      cwd: "../packages/backend",
-      env: {
-        ...process.env,
-        PATH: pathWithRepoBin,
-        NODE_OPTIONS: "--no-node-snapshot",
-      },
-      url: backendReadiness,
-      reuseExistingServer: !isCI,
-      timeout: 180 * 1000,
-      stdout: "pipe",
-      stderr: "pipe",
-    },
-    {
-      command: `janus-cli package start ${sharedConfigArgs}`,
+      command: `janus-cli package start ${harnessConfigArgs()}`,
       cwd: "../packages/app",
       env: { ...process.env, PATH: pathWithRepoBin },
       url: frontendUrl,
