+++ e2e-tests/playwright/support/local-harness-servers.ts
@@ -1,0 +1,58 @@
+/**
+ * Shared pieces of the cluster-free harness Playwright configs
+ * (playwright.legacy-local.config.ts and playwright.plugin-sanity.config.ts).
+ * Both boot `packages/backend` from source with the same config stack; only
+ * the frontend dev server and per-config overlays differ.
+ */
+
+import { resolve } from "path";
+
+import type { PlaywrightTestConfig } from "@playwright/test";
+
+export const backendUrl = "http://localhost:7007";
+const backendReadiness = `${backendUrl}/.backstage/health/v1/readiness`;
+export const isCI = process.env.CI !== undefined && process.env.CI !== "";
+
+// backstage-cli / janus-cli live in the repo-root node_modules/.bin, which
+// yarn does not surface for this workspace, so the CLIs are invoked with the
+// root .bin prepended to PATH and run from their package directory.
+const repoRootBin = resolve(process.cwd(), "..", "node_modules", ".bin");
+export const pathWithRepoBin = `${repoRootBin}:${process.env.PATH ?? ""}`;
+
+/**
+ * The shared app-config stack. The dynamic-plugins-root entry is generated by
+ * the local-harness populate scripts: it carries the pluginConfig blocks
+ * merged by install-dynamic-plugins exactly like the production container
+ * does, and must come after the static app-config.dynamic-plugins.yaml so its
+ * plugin config wins. Per-config overlays passed via `extraConfigs` land last
+ * and win over everything.
+ */
+export const harnessConfigArgs = (extraConfigs: string[] = []): string =>
+  [
+    "--config ../../app-config.yaml",
+    "--config ../../app-config.dynamic-plugins.yaml",
+    "--config ../../app-config.local-e2e.yaml",
+    "--config ../../dynamic-plugins-root/app-config.dynamic-plugins.yaml",
+    ...extraConfigs.map((config) => `--config ${config}`),
+  ].join(" ");
+
+type WebServer = Extract<NonNullable<PlaywrightTestConfig["webServer"]>, unknown[]>[number];
+
+/**
+ * webServer entry booting the RHDH backend from source. The command mirrors
+ * packages/backend's `start` script (--require instrumentation) — keep in sync.
+ */
+export const backendWebServer = (extraConfigs: string[] = []): WebServer => ({
+  command: `backstage-cli package start --require ./src/instrumentation.js ${harnessConfigArgs(extraConfigs)}`,
+  cwd: "../packages/backend",
+  env: {
+    ...process.env,
+    PATH: pathWithRepoBin,
+    NODE_OPTIONS: "--no-node-snapshot",
+  },
+  url: backendReadiness,
+  reuseExistingServer: !isCI,
+  timeout: 180 * 1000,
+  stdout: "pipe",
+  stderr: "pipe",
+});
